requirements.txt
ipykernel
scanpy
shiny
jinja2
anndata2ri